+++ lib/navbar/explore/explore.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:ui';
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -113,96 +112,84 @@ class _ExploreWordsMobileState extends ConsumerState<ExploreWordsMobile> {
               ),
             );
           }
-          final colorScheme = Theme.of(context).colorScheme;
           final words = request.data as List<Word>;
-          return Stack(
-            children: [
-              CustomPaint(
-                painter: BackgroundPainter(
-                  primaryColor: colorScheme.primary,
-                  secondaryColor: colorScheme.onPrimaryContainer,
-                ),
-              ),
-              BackdropFilter(filter: ImageFilter.blur(sigmaX: 60, sigmaY: 60), child: Container()),
-              Material(
-                color: Colors.transparent,
-                child: RefreshIndicator(
-                  onRefresh: () async {
-                    await exploreWords();
-                  },
-                  child: Stack(
-                    fit: StackFit.expand,
+          return Material(
+            color: Colors.transparent,
+            child: RefreshIndicator(
+              onRefresh: () async {
+                await exploreWords();
+              },
+              child: Stack(
+                fit: StackFit.expand,
+                children: [
+                  Column(
                     children: [
-                      Column(
-                        children: [
-                          Expanded(
-                            child: PageView.builder(
-                                itemCount: words.length,
-                                controller: exploreController.pageController,
-                                scrollBehavior: MaterialScrollBehavior(),
-                                onPageChanged: (x) {
-                                  // if (x > max - 5) {
-                                  //   page++;
-                                  //   exploreWords();
-                                  // }
-                                  if (x % _scrollCountCallback == 0) {
-                                    widget.onScrollThresholdReached!();
-                                  }
-                                  NavbarNotifier.hideSnackBar(context);
-                                },
-                                physics: ClampingScrollPhysics(),
-                                scrollDirection: Axis.vertical,
-                                itemBuilder: (context, index) {
-                                  return ExploreWord(word: words[index], index: index);
-                                }),
-                          ),
-                          AnimatedBuilder(
-                              animation: exploreController,
-                              builder: (context, child) {
-                                return exploreController.isAnimating
-                                    ? Container(
-                                        alignment: Alignment.center,
-                                        margin: EdgeInsets.only(
-                                          bottom: kBottomNavigationBarHeight + 50,
-                                        ),
-                                        // left: 120,
-                                        child: Column(
-                                          mainAxisAlignment: MainAxisAlignment.center,
-                                          children: [
-                                            Icon(
-                                              CupertinoIcons.chevron_up,
-                                              size: 30,
-                                            ),
-                                            Text(
-                                              'Swipe up to see more',
-                                              style: TextStyle(
-                                                fontSize: 16,
-                                                fontWeight: FontWeight.w500,
-                                              ),
-                                            ),
-                                          ],
-                                        ))
-                                    : SizedBox.shrink();
-                              })
-                        ],
+                      Expanded(
+                        child: PageView.builder(
+                            itemCount: words.length,
+                            controller: exploreController.pageController,
+                            scrollBehavior: MaterialScrollBehavior(),
+                            onPageChanged: (x) {
+                              // if (x > max - 5) {
+                              //   page++;
+                              //   exploreWords();
+                              // }
+                              if (x % _scrollCountCallback == 0) {
+                                widget.onScrollThresholdReached!();
+                              }
+                              NavbarNotifier.hideSnackBar(context);
+                            },
+                            physics: ClampingScrollPhysics(),
+                            scrollDirection: Axis.vertical,
+                            itemBuilder: (context, index) {
+                              return ExploreWord(word: words[index], index: index);
+                            }),
                       ),
-                      request.state == RequestState.active
-                          ? Positioned(
-                              bottom: kBottomNavigationBarHeight + 50,
-                              left: 120,
-                              child: Text(
-                                'Fetching more words',
-                                style: TextStyle(
-                                    fontSize: 16,
-                                    fontWeight: FontWeight.w500,
-                                    color: Colors.grey.shade600),
-                              ))
-                          : SizedBox.shrink(),
+                      AnimatedBuilder(
+                          animation: exploreController,
+                          builder: (context, child) {
+                            return exploreController.isAnimating
+                                ? Container(
+                                    alignment: Alignment.center,
+                                    margin: EdgeInsets.only(
+                                      bottom: kBottomNavigationBarHeight + 50,
+                                    ),
+                                    // left: 120,
+                                    child: Column(
+                                      mainAxisAlignment: MainAxisAlignment.center,
+                                      children: [
+                                        Icon(
+                                          CupertinoIcons.chevron_up,
+                                          size: 30,
+                                        ),
+                                        Text(
+                                          'Swipe up to see more',
+                                          style: TextStyle(
+                                            fontSize: 16,
+                                            fontWeight: FontWeight.w500,
+                                          ),
+                                        ),
+                                      ],
+                                    ))
+                                : SizedBox.shrink();
+                          })
                     ],
                   ),
-                ),
+                  request.state == RequestState.active
+                      ? Positioned(
+                          bottom: kBottomNavigationBarHeight + 50,
+                          left: 120,
+                          child: Text(
+                            'Fetching more words',
+                            style: TextStyle(
+                                fontSize: 16,
+                                fontWeight: FontWeight.w500,
+                                color: Colors.grey.shade600),
+                          ))
+                      : SizedBox.shrink(),
+                ],
               ),
-            ],
+            ),
           );
         });
   }
@@ -278,73 +265,6 @@ class _ExploreWordsDesktopState extends State<ExploreWordsDesktop> {
       ),
     );
   }
-}
-
-class BackgroundPainter extends CustomPainter {
-  final Color primaryColor;
-  final Color secondaryColor;
-
-  BackgroundPainter({this.primaryColor = Colors.green, this.secondaryColor = Colors.white});
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final paint = Paint();
-    paint.color = primaryColor;
-
-    paint.maskFilter = MaskFilter.blur(BlurStyle.normal, 200);
-    // paint square
-    canvas.drawRect(
-        Rect.fromCenter(
-          center: Offset(150, 200),
-          width: 200,
-          height: 200,
-        ),
-        paint);
-    canvas.translate(-180, -500);
-    // paint abstract shape
-    paint.color = secondaryColor;
-    canvas.drawPath(
-        Path()
-          ..moveTo(400, 500)
-          ..quadraticBezierTo(500, 400, 600, 500)
-          ..quadraticBezierTo(700, 600, 600, 700)
-          ..quadraticBezierTo(500, 800, 400, 700)
-          ..quadraticBezierTo(300, 600, 400, 500)
-          ..close(),
-        paint);
-    canvas.restore();
-    paint.blendMode = BlendMode.overlay;
-
-    // paint triangle
-    paint.color = Colors.blueAccent;
-    canvas.drawPath(
-        Path()
-          ..moveTo(-150, 400)
-          ..lineTo(200, 400)
-          ..lineTo(200, 600)
-          ..close(),
-        paint);
-
-    paint.color = Colors.orange.shade600;
-    canvas.drawPath(
-        Path()
-          ..moveTo(300, 400)
-          ..lineTo(600, 600)
-          ..lineTo(100, 700)
-          ..close(),
-        paint);
-
-    paint.color = Colors.redAccent;
-    canvas.drawCircle(Offset(200, 300), 100, paint);
-    paint.color = Colors.blue.shade600;
-
-    // / shift the canvas to the left
-
-    // add overlay blend mode
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) => false;
 }
 
 class ExploreWord extends ConsumerStatefulWidget {
